+++ docs/changelog.md
@@ -5,6 +5,7 @@ Versions and bullets are arranged chronologically from latest to oldest.
 ## v6.0.0
 
 - \[API\] Rename some Animator APIs.
+- \[fix\] Infinite duration check for reverse
 
 ## v5.0.0